chore(media_types): move `extensions` utility
