+++ media_types/_util.ts
@@ -13,9 +13,6 @@ export interface DBEntry {
   extensions?: string[];
 }
 
-/** A map of extensions for a given media type. */
-export const extensions: Map<string, string[]> = new Map();
-
 export function consumeToken(v: string): [token: string, rest: string] {
   const notPos = indexOf(v, isNotTokenChar);
   if (notPos === -1) {
