+++ media_types/extensions_by_type.ts
@@ -2,7 +2,7 @@
 // This module is browser compatible.
 
 import { parseMediaType } from "./parse_media_type.ts";
-import { extensions } from "./_util.ts";
+import { extensions } from "./_db.ts";
 import "./_db.ts";
 
 export { extensions };
